pyproject.toml
[project]
name = "challenge-airbnb"
version = "0.1.0"
description = ""
authors = [
    {name = "unatic",email = "38244346+unatic@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
